No need to collapse 3d to 2d vars for output
Previously, all of the multidimensional cosp output was collapsed by
combining dimensions into a mixed dimension for all fields that had more
than just a column and vertical dimension, for example,
clMISR(ncol,tau,cth) would be collapsed into something like
clMISR(ncol,tau*cth). I think this is because `cam\_history` used to not
support 3D output arrays, but this is no longer the case so we save a
lot of data copying and redundant variables and code by removing this.
Still BFB.