Fixes to KPP/CVMix
convective diffusivity was being applied one layer too low (essentially
a gap existed between the end of the boundary layer and the start of the
convective diffusivity.

simple shapes diffusivity was also not being added back in the boundary
layer as it should, this is fixed

interfaces for matchBoth are also fixed.  Previously setting that option
led to model crashes.